re-shuffle file structure
